Scroll to first error on invalid form submission
React-collapsible seems a bit buggy on how it handles its internal state
when triggered via the open prop. When trying to close or re-open the
collapsible after it has been opened the first time things don't quite
work right. For this prototype this is probably fine, but still not
perfect.

OCRVS-118