Allow secrets to be rotated properly
We cannot remove secrets from running containers. This means that for
every deploy we need to create new secrets with a different name and
update the services to use those new secrets. We will get a build up of
secrets over time as deploys are done, however these can be cleaned up
manually if necessary. If may be useful to keep these around as a
history.

OCRVS-319