Fixes verification code resend for demo users
Previously, if user asked to resend verification code after login,
the system would generate a new verification code. But this code
would never be sent to the user in a non-production environment,
making demo/testing difficult. This commit changes the behaviour
of the system to not generate a new verification code if the user
is has a "demo" role or the environment is non-production.

Should resovle Issue #127 in github.