Add new single root handler
The workflow service should accept FHIR-style requests only. We need to
remove the custom endpoint we are using and rather detect the event that
is happening from the data received in the FHIR request rather than by
the endpont used. This is so the service remains standards based for 3rd
party users.

We are temporarily keeping the old routes for reference.

OCRVS-918