Change locations and facilites store offline into a map
We are loading thousands of facilities and locations, it becomes slow to
filter an array of those to get a particular one by id particularly as
this is done multiple times per form render. This change migrates those
arrays to a maps with the id as a key. This improves lookup performance.

OCRVS-1346