Fix DOB validation (cannot be today)
Prior to this change, while registering a birth event, applicant could input the date of application creation as the birth date of mother and/or father.

This change will enforce the applicant to use a date from the past for the DOB of mother and father.
Additionally, this validation is also added to applicant's DOB while registering a death event.

Fixes OCRVS-1453