Fix query to show correct number of registrations
Prior to this commit, the query was not asking for the total number of
registrations that are accepted, hence the total number could not be
shown on the tab header.

This change will fix the query, so that it asks for the total number of
registrations that can be shown in the tab header.