Change expansion button from plus-minus to up-down
Prior to this change, the expansion buttons on 'Ready to print' tab
showed plus and minus buttons, like in all other tabs. It was
incompatible with the app's design.

This change will fix the icon issue by adding an optional prop to the
ExpansionButton element to decide the icon. It would be passed in
GridTable > ListItemAction > ExpansionButton order.