Add "Ready to Print" tab to RegistrarHome
Prior to this change, there was no way to show the applications that
are accepted by the registrar and ready to print.

This change adds a tab to registrar's home page, where registrar can
view the registered applications and print the certificates. The GQL
query and related data structure for the transformer function was also
modified to display contact number.