Fixes navigation from draft to in-progress tab
Prior to this change, when the user clicked 'Save & Exit'
while writing a draft application, the user was being
taken to review tab. It is supposed to take the user to
progress tab instead.

This change fixes this issue by checking the token payload
by token and examining the scope of the user.

Issues fixed by this commit: OCRVS-1746, OCRVS-1747