Fixes tests to increase code coverage
Prior to this change, the existing tests
were not checking whether pressing the
'Save & Exit' button was navigating to the
proper progress tab.

This change fixes those tests by providing
tokens for both field-agent and registrar
to test if the user is being directed to the
home page with progress tab selected.