Fix implementations of navigatioon functions
Prior to this change, goToHome function was
implemented in such a way that no matter which
tab the user was in, the user always would have
been taken to in-progress tab when save-&-exit
button was clicked.

This change fixes the issue so that the user
would be taken to in-progress tab only when the
user clicks save-&-exit button from an in-progress
draft. A new function has been added to do the
job. The goToHome function has been reverted to
the original version.