Unify the way we send GraphQL query errors to Sentry
The query error handling was a bit all over the place, so I created a new
wrapper component for Query, that handles the error cases automatically.

In the future this could also be used as the base for a more refined Query component,
that would also provide the correct return types based on the submitted query.