Show validation errors for unvisited group
Prior to this change:

We showed all validation errors for the fields of a visited group. But in
case of an unvisited group of partial data, the formik fields had to be
touched again to show errors.

After this change:

We still show all validation errors for the visited group but in
addition, a check has been added if the group is unvisited whether there
are any updated fields or not. If found, then make those formik fields
touched to show validation errors.