Make field details, form values and declaration values available for option conditional
We encountered a use case for only showing certain select options to specific user roles but only when the declaration is in DECLARED state.

This is a breaking change if anyone had been using the old 'optionCondition' as previously the function only got one parameter 'field'. Using a object parameter in this case however, ensures we can supply the function with richer data while being backwards compatible