fix bug causing a task without a 'reason' field ending up to search & metrics
This issuehappened randomly when a correction request was approved. The likely root cause was both tasks going into updateFHIRBundle where the lastModified time of the correction approving task got changed to either equal or greater value than what the proceeding 'corrected' task had. Later in the code when the latest task was selected to be sent to search and metrics as part of the record, the wrong task ended up being sent