feat: Update the order of system user add/edit form fields and labels and remove NID field  (#7697)
* Ask surname first in the User details form

we often receive requests from countries to have the surname first in this form, we have decided to satisfy that requirement by putting surname first for ALL countries.

https://github.com/opencrvs/opencrvs-core/issues/6830

* Change name and surname labels for User Details form

A request came for the clients to change the labels for "First name(s)" to "User's first name". Also change "Last name" to "User's surname" for clarity and less confussion.

https://github.com/opencrvs/opencrvs-core/issues/6830

* Remove NID field from the user details form

Remove the question that asks the user for a national ID number in this form as it is not required

https://github.com/opencrvs/opencrvs-core/issues/6830

* Remove lang suffix for FamilyName field

We no longer need the suffixes for the FamilyName field since we not really using it anywhere

https://github.com/opencrvs/opencrvs-core/issues/6830

* Add migration to remove NID field from user

A requirement from client to remove the NID field since most of the contries are not really using it.

https://github.com/opencrvs/opencrvs-core/issues/6830

* Record changes in the CHANGELOG

To keep track of changes done on this PR

https://github.com/opencrvs/opencrvs-core/issues/6830

* Remove Eng suffix for the firstName in user form

We currently don't support milti language support for both surname and
first names so the ENG/EN suffixes are not needed. Also remove EN bits
that were left for the surname field

https://github.com/opencrvs/opencrvs-core/issues/6830

---------

Co-authored-by: Siyasanga Mtshokotsha <siya@Siyasangas-MBP.lan>
Co-authored-by: Siyasanga Mtshokotsha <siya@Siyasangas-MacBook-Pro.local>