fix: use closest role entry when none available
Records generated for demo purposes include history entries dated
before even a user was created in the system. Therefore, those
history entries are unable to be matched against the role of the
user in that particular time period. We are now using the first
role entry available for that user in those cases.