fix: use earliest role entry available when none found at the given point in time (#9308)
* fix: use closest role entry when none available

Records generated for demo purposes include history entries dated
before even a user was created in the system. Therefore, those
history entries are unable to be matched against the role of the
user in that particular time period. We are now using the first
role entry available for that user in those cases.

* test: add test cases for getUserRoleFromHistory

* docs: add CHANGELOG entry

* chore: grammatical soundness

* docs: move CHANGELOG entry under 1.7.1