fix: switch to `classic` analyzer for name fields
Removing the custom `human_name` analyzer as the `word_delimiter`
filter was causing more issues than it solved. The classic
tokenizer is a grammar based tokenizer that is good for English
language documents. This tokenizer has heuristics for special
treatment of acronyms, company names, email addresses, and
internet host names.