feat: introduce typescript references in commons
The use of project references in typescript enables two things:
- The first one is the use of different configs for test and
implementation files. This enables us to not typecheck the test
files when any of the implementation changes, but still do so for
the test files when any of the implementation changes. It also
improved the startup time of tsserver in the editor.
- The other is to use a single `tsc -b` process to build both the
commonjs & esm builds. That way we can take advantage of incremental
builds which we were missing out on previously when using two
separate processes for it.