fix: keep `ActionProcedure` as an internal type (#10376)
Typescript tries to use the `ActionProcedure` type for building
the `AppRouter` one but due to how our build process works, the
type is not available to be imported by default. It will instead
need to be manually copied over from events/build to
toolkit/build. So we opted to not expose the type altogether.