fix: stricter dedup check
If the new event does not have data for a field that is being
reference in a dedup check clause, the clause will now be
considered as not matched. Previously it was taken considered as
matched instead which was causing many false negatives to be
flagged. This will need to be revisited when we introduce a
threshold value for matching duplicates.