fix: stricter dedup check (#10389)
* fix: stricter dedup check

If the new event does not have data for a field that is being
reference in a dedup check clause, the clause will now be
considered as not matched. Previously it was taken considered as
matched instead which was causing many false negatives to be
flagged. This will need to be revisited when we introduce a
threshold value for matching duplicates.

* test: generate applicant.dob

* test: populate `applicant.dob` in test data for duplicates

* fix: typo

* fix: return null if all the sub-queries of `or` are null

* test: should not find duplicates if values are missing

* test: provide all the values referenced in config