Add annotation field to the $lookup helper's resolved actions
The $action helper returns the action object for the PRINT_CERTIFICATE action. However, when $lookup tried to resolve properties from that action object, it was creating a resolvedAction object that didn't include the annotation field, even though annotation is a valid property of all actions (it's defined in the ActionBase schema).