fix: provide empty declaration object in `DUPLICATE_DETECTED` action
- Reason: when an event is requested on declare action, the declaration object has
verified status `pending`. Then the request is forwared to countryconfig and there
the verification takes place, so the field gets updated value as 'verified' or 'failed'
etc. But after that, if there are duplicates found, then sending the input declaration
(which had verified 'pending') again overwrites newly updated verification status